ARM64: dts: tegra210: remove CC7 state from CPU idle states
Because the CL-DVFS and BPMP-L (when doing CC7) are controlling CPU PMIC
through the same I2C bus, which means they are not able to work at the
same time. Only one can own the bus during the same period of time.

And we are using CL-DVFS on all CPU frequency range, which means we are
not able to reach CC7 during runtime. So removing it.

Bug 200277498

Change-Id: I52d535e9aaf656fb7ceb03e0f3ea3678afc0d171
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://git-master/r/1516070
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
(cherry picked from commit b32bcff022a449e8446ae6455fd3d20456885b31)